fix build issues and add data pages
