+++ src/app/demons/page.tsx
@@ -1,5 +1,5 @@
+import DemonCard from '../components/DemonCard';
 import { demons } from '../../data/demons';
-import DemonCard from '../../components/DemonCard';
 
 export default function DemonsPage() {
   return (
@@ -13,26 +13,4 @@ export default function DemonsPage() {
     </div>
   );
 }
-export const demons = [
-  {
-    id: 'lucifer',
-    name: 'Lucifer',
-    description: 'Kritęs angelas, vadinamas Ryto Žvaigžde.',
-    powers: ['Vilionė', 'Šviesos manipuliacija'],
-    domain: 'Maištas'
-  },
-  {
-    id: 'belial',
-    name: 'Belial',
-    description: 'Apgaulės ir chaoso valdovas.',
-    powers: ['Apgaulė', 'Manipuliacija'],
-    domain: 'Chaosas'
-  },
-  {
-    id: 'asmodeus',
-    name: 'Asmodeus',
-    description: 'Destrukcijos ir geismo demonas.',
-    powers: ['Kerštas', 'Geismo skatinimas'],
-    domain: 'Destrukcija'
-  }
-];
+
