+++ src/data/archangels.ts
@@ -1,0 +1,23 @@
+export const archangels = [
+  {
+    id: 'michael',
+    name: 'Michael',
+    description: 'Dangaus gynėjas ir karys.',
+    powers: ['Apsauga', 'Karinė strategija'],
+    domain: 'Gynyba'
+  },
+  {
+    id: 'gabriel',
+    name: 'Gabriel',
+    description: 'Dievo pasiuntinys.',
+    powers: ['Pranašystės', 'Žinios perdavimas'],
+    domain: 'Komunikacija'
+  },
+  {
+    id: 'raphael',
+    name: 'Raphael',
+    description: 'Sielų gydytojas.',
+    powers: ['Gydymas', 'Sielos apsauga'],
+    domain: 'Gydymas'
+  }
+];
