+++ src/data/demons.ts
@@ -1,0 +1,23 @@
+export const demons = [
+  {
+    id: 'lucifer',
+    name: 'Lucifer',
+    description: 'Kritęs angelas, vadinamas Ryto Žvaigžde.',
+    powers: ['Vilionė', 'Šviesos manipuliacija'],
+    domain: 'Maištas'
+  },
+  {
+    id: 'belial',
+    name: 'Belial',
+    description: 'Apgaulės ir chaoso valdovas.',
+    powers: ['Apgaulė', 'Manipuliacija'],
+    domain: 'Chaosas'
+  },
+  {
+    id: 'asmodeus',
+    name: 'Asmodeus',
+    description: 'Destrukcijos ir geismo demonas.',
+    powers: ['Kerštas', 'Geismo skatinimas'],
+    domain: 'Destrukcija'
+  }
+];
